refactor(similarity-prep): split orchestration modules
