+++ src/egui_app/controller/similarity_prep/mod.rs
@@ -1,7 +1,11 @@
 mod db;
 mod orchestration;
+mod plan;
 mod progress;
+mod runtime;
 mod state;
 mod store;
+mod transitions;
+mod ui;
 
 pub(super) const DEFAULT_CLUSTER_MIN_SIZE: usize = 10;
